requirements.txt
scapy==2.5.0
cryptography==41.0.7
pycryptodome==3.19.0
tkinter-tooltip==2.1.0